+++ .pnp.cjs
@@ -816,6 +816,10 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:packages/sources/view-function-multi-chain"\
       },\
       {\
+        "name": "@chainlink/view-starknet-latest-answer-adapter",\
+        "reference": "workspace:packages/sources/view-starknet-latest-answer"\
+      },\
+      {\
         "name": "@chainlink/wbtc-address-set-adapter",\
         "reference": "workspace:packages/sources/wbtc-address-set"\
       },\
@@ -1059,6 +1063,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@chainlink/vesper-adapter", ["workspace:packages/composites/vesper"]],\
       ["@chainlink/view-function-adapter", ["workspace:packages/sources/view-function"]],\
       ["@chainlink/view-function-multi-chain-adapter", ["workspace:packages/sources/view-function-multi-chain"]],\
+      ["@chainlink/view-starknet-latest-answer-adapter", ["workspace:packages/sources/view-starknet-latest-answer"]],\
       ["@chainlink/wbtc-address-set-adapter", ["workspace:packages/sources/wbtc-address-set"]],\
       ["@chainlink/wootrade-adapter", ["workspace:packages/sources/wootrade"]],\
       ["@chainlink/wrapped-adapter", ["workspace:packages/sources/wrapped"]],\
@@ -8716,6 +8721,22 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@types/node", "npm:16.18.96"],\
             ["ethers", "npm:5.7.2"],\
             ["nock", "npm:13.5.4"],\
+            ["tslib", "npm:2.4.1"],\
+            ["typescript", "patch:typescript@npm%3A5.0.4#~builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
+          ],\
+          "linkType": "SOFT"\
+        }]\
+      ]],\
+      ["@chainlink/view-starknet-latest-answer-adapter", [\
+        ["workspace:packages/sources/view-starknet-latest-answer", {\
+          "packageLocation": "./packages/sources/view-starknet-latest-answer/",\
+          "packageDependencies": [\
+            ["@chainlink/view-starknet-latest-answer-adapter", "workspace:packages/sources/view-starknet-latest-answer"],\
+            ["@chainlink/external-adapter-framework", "npm:0.33.8"],\
+            ["@types/jest", "npm:27.5.2"],\
+            ["@types/node", "npm:16.18.96"],\
+            ["nock", "npm:14.0.0-beta.6"],\
+            ["starknet", "npm:6.9.0"],\
             ["tslib", "npm:2.4.1"],\
             ["typescript", "patch:typescript@npm%3A5.0.4#~builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
           ],\