+++ packages/sources/view-starknet-latest-answer/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export { endpoint as functionEndpoint } from './function'